Subclass Clearance::UsersController as per docs
https://github.com/thoughtbot/clearance#controllers
