+++ config/initializers/clearance.rb
@@ -1,4 +1,5 @@
 Clearance.configure do |config|
   config.mailer_sender          = 'reply@example.com'
   config.rotate_csrf_on_sign_in = true
+  config.routes                 = false
 end
