--- config/initializers/clearance/params.rb
@@ -1,22 +1,0 @@
-module UserStrongParams
-
-  PARAMS = %i(first_name last_name email password)
-
-  def user_params
-    if creating? then user_strong_params else Hash.new end
-  end
-
-  private
-
-  def creating?
-    action_name.to_sym.eql? :create
-  end
-
-  def user_strong_params
-    params.require(
-      Clearance.configuration.user_parameter
-    ).permit(PARAMS)
-  end
-end
-
-Clearance::UsersController.send(:prepend, UserStrongParams)
