+++ config/routes.rb
@@ -2,6 +2,19 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   root to: 'application#welcome'
 
+  resources :passwords, controller: 'clearance/passwords', only: %i(create new)
+  resource :session, controller: 'clearance/sessions', only: %i(create)
+
+  resources :users, controller: 'users', only: %i(create) do
+    resource :password,
+             controller: 'clearance/passwords',
+             only: %i(create edit update)
+  end
+
+  get '/sign_in' => 'clearance/sessions#new', as: 'sign_in'
+  delete '/sign_out' => 'clearance/sessions#destroy', as: 'sign_out'
+  get '/sign_up' => 'clearance/users#new', as: 'sign_up'
+
   resource :profile, only: %i(show update)
 
   namespace :order do
